FACT-59: Update unit tests.
