+++ src/main/views/choose-service.njk
@@ -20,7 +20,7 @@
 
 {% block content %}
 {% if errors %}
-  {{ errorSummary(error.title, error.text, "#choose-service-error") }}
+  {{ errorSummary(error.title, error.text, "#radio-buttons") }}
 {% endif %}
 
 <div class="govuk-grid-row">
@@ -43,9 +43,9 @@
             </span>
         {% endif %}
         {% for result in results %}
-            <div class="govuk-radios__item">
-                <input class="govuk-radios__input" id={{ result.name | lower | replace(" ", "-") | replace(",", "") }} name="chooseService" type="radio" value={{ result.name | lower | replace(" ", "-") | replace(",", "") }}>
-                <label class="govuk-label govuk-radios__label govuk-label--s" for={{ result.name | lower | replace(" ", "-") | replace(",", "") }}>
+            <div class="govuk-radios__item" id="radio-buttons">
+                <input class="govuk-radios__input" id={{ result.name }} name="chooseService" type="radio" value={{ result.name | lower | replace(" ", "-") | replace(",", "") }}>
+                <label class="govuk-label govuk-radios__label govuk-label--s" for={{ result.name }}>
                     {{ result.name }}
                 </label>
                 <div id={{ result.description }} class="govuk-hint govuk-radios__hint">
